requirements.txt
charset-normalizer==3.4.1
gitdb==4.0.12
GitPython==3.1.44
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logging==0.4.9.6
matplotlib==3.10.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
plotly==6.0.1
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.15.2
SQLAlchemy==2.0.37
streamlit==1.43.1
urllib3==2.3.0
